requirements.txt
numpy>=1.24
pandas>=1.5
PyYAML>=6.0